requirements.txt
geographiclib==1.50
geopy==2.1.0
Pillow==8.2.0
